TileMover parent class
Changed Unit to inherit from a parent class 'TileMover'.
TileMover handles basic moving of Node2Ds around the map, allowing the UnitController to correctly place them on any Tile with consideration for the TopLayer and its YOffsets.
Formation UI control buttons can now be or inherit from TileMover to be able to be moved into position on top of a formation by the UnitController.